requirements.txt
python-telegram-bot==21.6
anthropic>=0.34.0
flask>=3.0.0
python-dotenv>=1.0.0
